+++ src/yamk/functions.py
@@ -34,6 +34,8 @@ class Exists(Function):
     name = "exists"
 
     def __call__(self, path):
+        if isinstance(path, list):
+            return [self(file) for file in path]
         return self.base_dir.joinpath(path).exists()
 
 
@@ -41,6 +43,8 @@ class Stem(Function):
     name = "stem"
 
     def __call__(self, path):
+        if isinstance(path, list):
+            return [self(file) for file in path]
         return self.base_dir.joinpath(path).stem
 
 
@@ -48,6 +52,8 @@ class Parent(Function):
     name = "parent"
 
     def __call__(self, path):
+        if isinstance(path, list):
+            return [self(file) for file in path]
         return self.base_dir.joinpath(path).parent